Consolidate things
