--- src/compiler/typescript_to_rust/azle_generate/src/ts_ast/ts_type/mod.rs
@@ -1,4 +1,0 @@
-mod ts_fn_or_constructor_type;
-mod ts_type;
-
-pub mod ts_type_ref;
