--- src/compiler/typescript_to_rust/azle_generate/src/ts_ast/ts_type/ts_fn_or_constructor_type.rs
@@ -1,14 +1,0 @@
-use swc_ecma_ast::TsFnOrConstructorType;
-
-use crate::traits::GetSpan;
-
-impl GetSpan for TsFnOrConstructorType {
-    fn get_span(&self) -> swc_common::Span {
-        match self {
-            TsFnOrConstructorType::TsFnType(ts_fn_type) => ts_fn_type.span,
-            TsFnOrConstructorType::TsConstructorType(ts_constructor_type) => {
-                ts_constructor_type.span
-            }
-        }
-    }
-}
