--- src/compiler/typescript_to_rust/azle_generate/src/ts_ast/ts_type/ts_type_ref.rs
@@ -1,14 +1,0 @@
-use swc_ecma_ast::{TsType, TsTypeRef};
-
-pub trait GetEnclosedTsTypes {
-    fn get_enclosed_ts_types(&self) -> Vec<TsType>;
-}
-
-impl GetEnclosedTsTypes for TsTypeRef {
-    fn get_enclosed_ts_types(&self) -> Vec<TsType> {
-        match &self.type_params {
-            Some(params) => params.params.iter().map(|param| *param.clone()).collect(),
-            None => vec![],
-        }
-    }
-}
